restructure project into multiple files
