+++ src/include/main.h
@@ -1,21 +1,23 @@
-#ifndef MAIN_H
-#define MAIN_H
+#ifndef __MAIN_H__
+#define __MAIN_H__
 
 // OS-dependent libs
 #ifdef _WIN32
     #include <ncurses/curses.h>
-    #include <stdio.h>
 #else
     #include <curses.h>
-#endif
+#endif // _WIN32
 
+// libs
+#include <stdio.h>
 #include <assert.h>
 #include <string.h>
 #include <stdlib.h>
 
+/* KEYS */
+
 // key defines
 #define ctrl(x) ((x) & 0x1f)
-
 #define ESCAPE 27
 #define ENTER 10
 #define TAB 9
@@ -32,13 +34,15 @@
 
 #define ESCDELAY 10
 
+/* TYPEDEFS */
+
 /* DATA */
 
 typedef struct ROW {
 	size_t index;
 	size_t size;
 	char* contents;
-} row_t;
+} row_t, *prow_t;
 
 typedef struct BUFFER {
 	row_t rows[MAX_ROWS];
@@ -46,6 +50,14 @@ typedef struct BUFFER {
 	size_t cur_pos;
 	size_t row_s;
 	char* filename;
-} buffer_t;
+} buffer_t, *pbuffer_t;
 
-#endif
+// local libs
+#include "input.h"
+#include "output.h"
+
+/* Functions */
+
+void read_file_to_buffer(buffer_t* buffer, char* filename);
+
+#endif // __MAIN_H__
