+++ src/include/output.h
@@ -1,0 +1,12 @@
+#ifndef __OUTPUT_H__
+#define __OUTPUT_H__
+
+#include "main.h"
+
+/* Draw buffer */
+void draw_buffer(buffer_t* buffer);
+
+/* Draw status bar */
+void draw_status_bar(buffer_t* buffer, int* ch);
+
+#endif
